Adding anchor points and smarter positioning.
You can now position a sprite using the  proprety, that
gets a tuple and uses the set anchor to modify the sprite's rect object.

The  property is also introduced, allowing you to alternate
between using any point inside the sprite as an anchor for moving
and modifying the sprite object.