Refactoring scale and rotate into image property.
First major refactor: `image` is now a property
with a getter that takes into account the values
passed to `scale` and `rotate`, and that's the
only place they're currently used (as opposed to
before, when transforms were made in their
setters).

Consequently, I'm facing a big issue here:
The `scale` and `rotate` attributes do not take
the `anchor` property into account for now.
Still a TODO.